set-progs增加组件开关控制；修复手误bug

+++ set-progs.sh
@@ -4,30 +4,20 @@ echo "=== running $(basename $0) ==="
 
 source config
 
-# ssh u@h 'cmd' 作为no-login shell无法执行source ~/.bashrc，出此下策
-JAVA_HOME=$libdir/jdk/"`ls $libdir/jdk`"
-HADOOP_HOME=$libdir/hadoop/"`ls $libdir/hadoop`"
-HBASE_HOME=$libdir/hbase/"`ls $libdir/hbase`"
-PHOENIX_HOME=$libdir/phoenix/"`ls $libdir/phoenix`"
-
 # source ~/.bashrc # 除在本地外无效
 
-# 复制phoenix server jar到hbase lib，此处可用ln -s
-sudo cp $PHOENIX_HOME/phoenix-*-HBase-*-server.jar $HBASE_HOME/lib/
+# ssh u@h 'cmd' 作为no-login shell无法执行source ~/.bashrc，出此下策
+JAVA_HOME=$libdir/jdk/"`ls $libdir/jdk`"
+# HADOOP_HOME=$libdir/hadoop/"`ls $libdir/hadoop`"
+# HBASE_HOME=$libdir/hbase/"`ls $libdir/hbase`"
+# PHOENIX_HOME=$libdir/phoenix/"`ls $libdir/phoenix`"
 
-# set_hadoop_env
-sed -i "s@.*export JAVA_HOME=.*@export JAVA_HOME=$JAVA_HOME@" $HADOOP_HOME/etc/hadoop/hadoop-env.sh
-
-# set_hadoop_slaves
-rm -f $HADOOP_HOME/etc/hadoop/slaves
-for slave in ${slaves[@]}; do
-cat >> $HADOOP_HOME/etc/hadoop/slaves << EOF
-$slave
-EOF
-done
-
+##### HADOOP #####
+if [[ -n `echo "${prog_arr[*]}" | grep jdk` ]]; then # 是否配置安装hadoop
+    HADOOP_HOME=$libdir/hadoop/"`ls $libdir/hadoop`"
+    if [ $? -eq 0 ]; then # 是否存在hadoop安装后目录
 # set_hadoop_core_site
-cat > $HADOOP_HOME/etc/hadoop/core-site.xml << EOF
+cat > $HADOOP_HOME/etc/hadoop/core-site.xml <<- EOF
 <?xml version="1.0"?>
 <?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
 <configuration>
@@ -69,7 +59,7 @@ cat > $HADOOP_HOME/etc/hadoop/core-site.xml << EOF
 EOF
 
 # set_hadoop_hdfs_site
-cat > $HADOOP_HOME/etc/hadoop/hdfs-site.xml << EOF
+cat > $HADOOP_HOME/etc/hadoop/hdfs-site.xml <<- EOF
 <?xml version="1.0"?>
 <?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
 <configuration>
@@ -98,27 +88,46 @@ cat > $HADOOP_HOME/etc/hadoop/hdfs-site.xml << EOF
 </configuration>
 EOF
 
-# set_hbase_env
-sed -i "s@.*export JAVA_HOME=.*@export JAVA_HOME=$JAVA_HOME@" $HBASE_HOME/conf/hbase-env.sh
-sed -i "s@.*export HBASE_PID_DIR=.*@export HBASE_PID_DIR=$tmpdir/hbase/pids@" $HBASE_HOME/conf/hbase-env.sh
-sed -i "s@.*export HBASE_MANAGES_ZK=.*@export HBASE_MANAGES_ZK=true@" $HBASE_HOME/conf/hbase-env.sh
-# export HBASE_CLASSPATH=$HBASE_HOME/conf
+        # set_hadoop_env
+        sed -i "s@.*export JAVA_HOME=.*@export JAVA_HOME=$JAVA_HOME@" $HADOOP_HOME/etc/hadoop/hadoop-env.sh
 
-# set_hbase_regionservers
-rm -f $HBASE_HOME/conf/regionservers
-for server in ${regionservers[@]}; do
-cat >> $HBASE_HOME/conf/regionservers << EOF
+        # set_hadoop_slaves
+        rm -f $HADOOP_HOME/etc/hadoop/slaves
+        for slave in ${slaves[@]}; do
+cat >> $HADOOP_HOME/etc/hadoop/slaves <<- EOF
+$slave
+EOF
+        done
+
+    fi
+fi
+
+##### HBASE #####
+if [[ -n `echo "${prog_arr[*]}" | grep hbase` ]]; then
+    HBASE_HOME=$libdir/hbase/`ls $libdir/hbase`
+    if [ $? -eq 0 ]; then
+
+        # set_hbase_env
+        sed -i "s@.*export JAVA_HOME=.*@export JAVA_HOME=$JAVA_HOME@" $HBASE_HOME/conf/hbase-env.sh
+        sed -i "s@.*export HBASE_PID_DIR=.*@export HBASE_PID_DIR=$tmpdir/hbase/pids@" $HBASE_HOME/conf/hbase-env.sh
+        sed -i "s@.*export HBASE_MANAGES_ZK=.*@export HBASE_MANAGES_ZK=true@" $HBASE_HOME/conf/hbase-env.sh
+        # export HBASE_CLASSPATH=$HBASE_HOME/conf
+
+        # set_hbase_regionservers
+        rm -f $HBASE_HOME/conf/regionservers
+        for server in ${regionservers[@]}; do
+cat >> $HBASE_HOME/conf/regionservers <<- EOF
 $server
 EOF
-done
+        done
 
-# set_hbase_site
-quorum_val=''
-for quorum in ${quorums[@]}; do
-	quorum_val="$quorum_val,$quorum"
-done
-quorum_val=${quorum_val#?}
-cat > $HBASE_HOME/conf/hbase-site.xml << EOF
+        # set_hbase_site
+        quorum_val=''
+        for quorum in ${quorums[@]}; do
+            quorum_val="$quorum_val,$quorum"
+        done
+        quorum_val=${quorum_val#?}
+cat > $HBASE_HOME/conf/hbase-site.xml <<- EOF
 <?xml version="1.0"?>
 <?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
 <configuration>
@@ -177,10 +186,22 @@ cat > $HBASE_HOME/conf/hbase-site.xml << EOF
 	</property>
 </configuration>
 EOF
+    fi
+fi
 
-# set_phoenix_hbase_site
-cp $HBASE_HOME/conf/hbase-site.xml $PHOENIX_HOME/bin/
+##### PHOENIX #####
+if [[ -n `echo "${prog_arr[*]}" | grep phoenix` ]]; then
+    PHOENIX_HOME=$libdir/phoenix/`ls $libdir/phoenix`
+    if [ $? -eq 0 ]; then
+        # set_phoenix_hbase_site
+        cp $HBASE_HOME/conf/hbase-site.xml $PHOENIX_HOME/bin/
 
+        # 复制phoenix server jar到hbase lib，此处可用ln -s
+        sudo cp $PHOENIX_HOME/phoenix-*-HBase-*-server.jar $HBASE_HOME/lib/
+    if
+fi
+
+##### unset #####
 # 避免和~/.bashrc变量混淆、冲突
 unset JAVA_HOME
 unset HADOOP_HOME
